+++ src/bpp/models/autor.py
@@ -67,6 +67,10 @@ class Autor(ModelZAdnotacjami, ModelZPBN_ID):
     imiona = models.CharField(max_length=512, db_index=True)
     nazwisko = models.CharField(max_length=256, db_index=True)
     tytul = models.ForeignKey(Tytul, CASCADE, blank=True, null=True)
+    pseudonim = models.CharField(max_length=300, blank=True, null=True, help_text="""
+    Jeżeli w bazie danych znajdują się autorzy o zbliżonych imionach, nazwiskach i tytułach naukowych,
+    skorzystaj z tego pola aby ułatwić ich rozróżnienie. Pseudonim pokaże się w polach wyszukiwania
+    oraz na podstronie autora, po nazwisku i tytule naukowym.""")
 
     aktualny = models.BooleanField(
         "Aktualny?",
@@ -174,6 +178,10 @@ class Autor(ModelZAdnotacjami, ModelZPBN_ID):
 
         if self.tytul is not None:
             buf += ", " + self.tytul.skrot
+
+        if self.pseudonim is not None:
+            buf += " (" + self.pseudonim + ")"
+
         return buf
 
     def dodaj_jednostke(self, jednostka, rok=None, funkcja=None):